Make battery optimizations prompt less confusing.
In the current state, the prompt reads:
> Disable Battery Optimizations
> 
> Do you really want to keep battery
> optimization switched on for Threema?
> Desktop/Web will not work correctly.
>
> No Yes

If the user only reads the title and the options, they are going to click the wrong one.

This problem is compounded by the fact that prompt comes right after a similar looking prompt in which the options are "Continue anyway" and "Disable".

This change modifies the title to be a question for which the options are answers that make sense.

An alternative fix would involve keeping the title as is and modify the content and the options. However that requires also modifying all the translations synchronously. With this change, translations can be modified independently.